requirements.txt
langchain
streamlit
langchain-openai
langchain-community
langchain-google-genai
langchainhub
brave-search
python-dotenv
chromadb
wikipedia